test: fix tests to work with restify@9 (#3031)
restify@9 was recently released. It only works with node >=v14.18.0
because it added use of https://nodejs.org/api/modules.html#core-modules
This updates restify testing to skip out when there is an incompat
(restify, node) version pair.

Also an unrelated change in "http-run-context.test.js" to get better
debug output if a spurious error is hit in it again.